react cleaner test
